Improve handling of missing files in XKB scripts

+++ files/xkb/install-xcompose.sh
@@ -7,6 +7,11 @@ layout=${1:-$layout}
 xcompose_file=${XCOMPOSEFILE:-$HOME/.XCompose}
 include_klfc="include \"%H/.XCompose-klfc-$layout\""
 
+if ! [ -f "$dir/XCompose" ]; then
+  "$dir/uninstall-xcompose.sh" "$layout"
+  exit
+fi
+
 if ! [ -f "$xcompose_file" ]; then
   echo 'include "%L"' > "$xcompose_file"
 fi
